IKN mappe tilføjet

+++ I4DAB/Handin2_1/Handin2_1/Telefon.cs
@@ -12,6 +12,6 @@ namespace Handin2_1
         public virtual string _TeleSelskab { get; set; }
         public virtual string _TelefonType { get; set; }
         public virtual string _Nummer { get; set; }
-        public virtual int _PersonID { get; set; }
+        public virtual long _PersonID { get; set; }
     }
 }
